+++ src/dinero.rs
@@ -3,6 +3,12 @@ use std::env;
 fn main() {
     match dinero::run_app(env::args().collect()) {
         Ok(_) => {}
-        Err(_) => std::process::exit(1),
+        Err(x) => {
+            let message = format!("{}", x);
+            if !message.is_empty() {
+                eprintln!("{}", message);
+            }
+            std::process::exit(1)
+        }
     }
 }