+++ src/parser.rs
@@ -7,9 +7,11 @@
 //! - Directive commodity
 
 use std::collections::HashSet;
+use std::convert::TryFrom;
 use std::fs::read_to_string;
 use std::path::PathBuf;
 
+use crate::error::MissingFileError;
 use crate::models::{Account, Comment, Currency, HasName, Payee, Transaction};
 use crate::parser::utils::count_decimals;
 use crate::{models, CommonOpts, List};
@@ -90,21 +92,25 @@ pub struct Tokenizer<'a> {
     seen_files: HashSet<&'a PathBuf>,
 }
 
-impl<'a> From<&'a PathBuf> for Tokenizer<'a> {
-    fn from(file: &'a PathBuf) -> Self {
+impl<'a> TryFrom<&'a PathBuf> for Tokenizer<'a> {
+    type Error = Box<dyn std::error::Error>;
+    fn try_from(file: &'a PathBuf) -> Result<Self, Self::Error> {
         match read_to_string(file) {
             Ok(content) => {
                 let mut seen_files: HashSet<&PathBuf> = HashSet::new();
                 seen_files.insert(file);
-                Tokenizer {
+                Ok(Tokenizer {
                     file: Some(file),
                     content,
                     seen_files,
-                }
+                })
             }
-            Err(err) => {
-                panic!("{:?}", err);
-            }
+            Err(err) => match err.kind() {
+                std::io::ErrorKind::NotFound => Err(Box::new(
+                    MissingFileError::JournalFileDoesNotExistError(file.to_path_buf()),
+                )),
+                _ => Err(Box::new(err)),
+            },
         }
     }
 }
@@ -148,7 +154,7 @@ impl<'a> Tokenizer<'a> {
                                 Rule::include => {
                                     // This is the special case
                                     let mut new_ledger =
-                                        self.include(inner, options, &ledger.commodities);
+                                        self.include(inner, options, &ledger.commodities).unwrap();
                                     ledger.append(&mut new_ledger);
                                 }
                                 Rule::price => {